tox -e circular: we need constrain_package_deps = true
Otherwise we get latest `plone.app.multilingual` 8 which is meant for Plone 6.1 and depends on `Products.CMFPlone`.
But we get `Products.CMFPlone` 6.0.10, which is the latest non-pre release, and this depends on `plone.app.multilingual`.
So we have a circular import.

By default we set `constrain_package_deps` to `true` in `plone/meta`, but for some reason when using `mxdev` we set it to `false`.
That gives the error here.
See also https://github.com/plone/meta/pull/220